Extract print methods to seperate classes
The printing methods are already pretty large, making it difficult to
add extra functionality. Extracting them to seperate classes allows
refactoring them for easier maintainability.

A lot of the functionality of calculating the terminal width can be
extracted to a separate object as well.
